// fontend/src/api/generated/model/productVariant.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * E-Commerce API
 * REST API for product management.

**Response envelope**
- Success: `{ success: true, data: <payload> }`
- Error: `{ success: false, error: '<message>' }`
 * OpenAPI spec version: 1.0.0
 */
import type { VariantOptionValue } from './variantOptionValue';

export interface ProductVariant {
  id: string;
  productId: string;
  /** @minLength 1 */
  sku: string;
  /** @exclusiveMinimum 0 */
  price: number;
  /**
     * @exclusiveMinimum 0
     * @nullable
     */
  comparePrice?: number | null;
  /** @minimum 0 */
  stock?: number;
  /**
     * @exclusiveMinimum 0
     * @nullable
     */
  weight?: number | null;
  /** @nullable */
  image?: string | null;
  isActive?: boolean;
  /** @nullable */
  createdAt: string | null;
  /** @nullable */
  updatedAt: string | null;
  variantOption: VariantOptionValue[];
}